sistema_encuestas_backend/requirements.txt
# Dependencias del Sistema de Encuestas con Recompensas
# =====================================================

# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Autenticación y seguridad
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Validación de datos
pydantic==2.5.0
email-validator==2.1.0

# CORS y middleware
fastapi-cors==0.0.6

# Fecha y tiempo
python-dateutil==2.8.2

# Archivos y uploads (futuro)
aiofiles==23.2.1
pillow==10.0.1


# HTTP client (para Google OAuth)
httpx==0.25.2

# Logging y monitoreo
structlog==23.2.0

# Variables de entorno
python-dotenv==1.0.0

# Rate limiting (futuro)
slowapi==0.1.9

# Caching (futuro)
redis==5.0.1

# Email (futuro)
fastapi-mail==1.4.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Desarrollo
black==23.11.0
isort==5.12.0
flake8==6.1.0 

uvicorn
aiosmtplib==2.0.2
jinja2==3.1.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1 